backend/em_store/requirements.txt
# Core Python
setuptools>=69.5.1

# Django and DB
Django==4.2.7
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.9.0
psycopg[binary]==3.2.9
dj-database-url==2.3.0

# REST API & Auth
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.0
dj-rest-auth==7.0.1
django-allauth==0.54.0

# File uploads & utils
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
cryptography==42.0.5

# API Documentation
drf-spectacular==0.28.0
drf-yasg==1.21.7

# Email
sendgrid==6.12.4

# Development tools
django-extensions==3.2.3
django-import-export==3.3.6

# File type detection
python-magic-bin==0.4.14; sys_platform == 'win32'
python-magic==0.4.27; sys_platform != 'win32'

# Cloud Storage
django-storages==1.14.2
boto3==1.34.0
botocore>=1.34.0
urllib3>=1.26.0
certifi>=2024.2.2

# Optional: for file/image handling (uncomment if needed)
# Pillow==10.3.0
# python-magic==0.4.27
